Feature Completion - Add new Registers and Moments
Now user can add new registers and moments to the circuit directly from the UI. The feature was left incomplete, which has been completed with a caveat. The caveat is that, all the circuits on the website has been named 'playground'. If someone could show how to dynamically pickup name of the circuit then it would make this feature truly complete